Refactor task queue inizialization
